requirements/base.txt
django==4.2.3
djangorestframework==3.14.0
drf-nested-routers==0.93.4
djoser==2.2.0
djangorestframework_simplejwt==5.2.2
django-filter==23.2
markdown==3.4.4